url_verification/requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
pycryptodome==3.19.1
requests==2.31.0 